requirements.txt
django-getpaid==2.0.0rc2
